add delete routes and buttons

+++ cookiecutter_mbam/experiment/views.py
@@ -5,17 +5,16 @@ import traceback
 from flask import Blueprint, render_template, request, flash, redirect, url_for, session
 from flask_security import current_user, login_required
 from cookiecutter_mbam.utils.error_utils import flash_errors
+from cookiecutter_mbam.utils.model_utils import resource_belongs_to_user
 from .forms import ExperimentForm, ExperimentAndScanForm, FlaskForm
 from .models import Experiment
 from .service import ExperimentService
 from cookiecutter_mbam.scan.service import ScanService
 from cookiecutter_mbam.base.tasks import global_error_handler
-
-
 from flask import current_app
+
 def debug():
     assert current_app.debug == False, "Don't panic! You're here by request of debug()"
-
 
 blueprint = Blueprint('experiment', __name__, url_prefix='/experiments', static_folder='../static')
 
@@ -31,13 +30,6 @@ def add_experiment(form, files):
     es.add(date=form.date.data, scanner=form.scanner.data, field_strength=form.field_strength.data,
                  user=current_user, files=files)
 
-@blueprint.route('/')
-def experiments():
-    """List experiments."""
-    # todo: this makes no sense, please eventually change.
-    experiments = Experiment.query.all()
-    return render_template('experiments/experiments.html', experiments=experiments)
-
 def scan_number_validation(request):
     """Validate that the number of scan files for a given experiment is at least one and no more than three"""
 
@@ -49,7 +41,6 @@ def scan_number_validation(request):
             return 'You can upload up to three files.'
     else:
         return ''
-
 
 @blueprint.route('/dev_add', methods=['GET', 'POST'])
 @login_required
@@ -63,9 +54,8 @@ def dev_add():
         return redirect(url_for('experiment.experiments'))
     return render_template('experiments/exp_and_scans.html', form=form)
 
-
-@login_required
 @blueprint.route('/add', methods=['GET', 'POST'])
+@login_required
 def add():
     """Access the add an experiment route and form."""
 
@@ -79,13 +69,12 @@ def add():
         if len(scan_number_error):
             flash(scan_number_error, 'warning')
             return redirect(url_for('experiment.add'))
-
         try:
             files = request.files.getlist('scan_file')
             add_experiment(form, files)
             num_scans = len(files)
 
-            flash('You successfully started the process of adding {}.'.format(num2words[num_scans]), 'success')
+            flash('You successfully started the process of adding {}. You should receive emails about upload status shortly.'.format(num2words[num_scans]), 'success')
 
         except Exception as e:
             flash('There was a problem uploading your scan', 'error')  # todo: error should be color coded red
@@ -93,45 +82,51 @@ def add():
                                  user_email=current_user.email, user_message='generic_message', email_user=True,
                                  email_admin=True)
 
-        return redirect(url_for('experiment.experiments'))
+        return redirect(url_for('public.home'))
     else:
         flash_errors(form)
+
     return render_template('experiments/experiment_and_scans.html',form=form)
 
-@blueprint.route('/<id>', methods=['GET'])
-def single_experiment(id):
-    """Display a single experiment"""
-    return render_template('experiments/experiment.html', id=id)
-
 @blueprint.route('/<id>/edit', methods=['GET', 'POST'])
+@login_required
 def edit_experiment(id):
     """Access and edit experiment metadata."""
-    exp = Experiment.query.filter(Experiment.id==id).first_or_404()
-    form = ExperimentForm(obj=exp)
+    if resource_belongs_to_user(Experiment, id):
+        exp = Experiment.query.filter(Experiment.id==id).first_or_404()
+        form = ExperimentForm(obj=exp)
 
-    if form.validate_on_submit():
-        form.populate_obj(exp) # update whatever has been changed in the form
-        exp.save()
-        flash('Experiment metadata updated','success')
-        return redirect(url_for('display.displays'))
+        if form.validate_on_submit():
+            form.populate_obj(exp) # update whatever has been changed in the form
+            exp.save()
+            flash('Experiment metadata updated','success')
+            return redirect(url_for('display.displays'))
+        else:
+            flash_errors(form)
+
+        return render_template('experiments/edit_experiment.html',session_form=form, experiment=exp)
     else:
-        flash_errors(form)
-
-    return render_template('experiments/edit_experiment.html',session_form=form, experiment=exp)
+        return render_template('403.html')
 
 @blueprint.route('/<id>/delete', methods=['GET', 'POST'])
+@login_required
 def delete_experiment(id):
     """Access and edit experiment metadata."""
-    exp = Experiment.query.filter(Experiment.id==id).first_or_404()
-    form = FlaskForm()
+    if resource_belongs_to_user(Experiment, id):
+        exp = Experiment.query.filter(Experiment.id==id).first_or_404()
+        form = FlaskForm()
 
-    if form.validate_on_submit():
-        exp.delete()
-        # for scan in exp.scans:
-        #     scan.delete()
-        flash('Deleted the session.','success')
-        return redirect(url_for('display.displays'))
+        if form.validate_on_submit():
+            for scan in exp.scans:
+                # Replace with ScanService.delete() to also delete from XNAT?
+                scan.delete()
+            exp.delete()
+
+            flash('Deleted the session.','success')
+            return redirect(url_for('display.displays'))
+        else:
+            flash_errors(form)
+
+        return render_template('experiments/delete_experiment.html', session_form=form, experiment=exp)
     else:
-        flash_errors(form)
-
-    return render_template('experiments/delete_experiment.html', session_form=form, experiment=exp)
+        return render_template('403.html')
